move `FeedImageLoaderWithFallbackComposite` to production
