+++ EssentialApp/EssentialApp/FeedImageLoaderWithFallbackComposite.swift
@@ -1,0 +1,42 @@
+//
+//  FeedImageLoaderWithFallbackComposite.swift
+//  EssentialApp
+//
+//  Created by Anthony on 13/7/25.
+//
+import Foundation
+import EssentialFeed
+
+public final class FeedImageLoaderWithFallbackComposite: FeedImageDataLoader {
+    private let primary: FeedImageDataLoader
+    private let fallback: FeedImageDataLoader
+    
+    public init(primary: FeedImageDataLoader, fallback: FeedImageDataLoader) {
+        self.primary = primary
+        self.fallback = fallback
+    }
+    
+    private class TaskWrapper: FeedImageDataLoaderTask {
+        var wrapped: FeedImageDataLoaderTask?
+        
+        func cancel() {
+            wrapped?.cancel()
+        }
+    }
+    
+    public func loadImageData(from url: URL, completion: @escaping (FeedImageDataLoader.Result) -> Void) -> FeedImageDataLoaderTask {
+        let task = TaskWrapper()
+        
+        task.wrapped = primary.loadImageData(from: url) { [weak self] primaryResult in
+            switch primaryResult {
+            case .success:
+                completion(primaryResult)
+                
+            case .failure:
+                task.wrapped = self?.fallback.loadImageData(from: url, completion: completion)
+            }
+        }
+        
+        return task
+    }
+}
