+++ EssentialApp/EssentialAppTests/FeedImageLoaderWithFallbackCompositeTests.swift
@@ -1,39 +1,6 @@
 import XCTest
 import EssentialFeed
-
-final class FeedImageLoaderWithFallbackComposite: FeedImageDataLoader {
-    private let primary: FeedImageDataLoader
-    private let fallback: FeedImageDataLoader
-    
-    init(primary: FeedImageDataLoader, fallback: FeedImageDataLoader) {
-        self.primary = primary
-        self.fallback = fallback
-    }
-    
-    private class TaskWrapper: FeedImageDataLoaderTask {
-        var wrapped: FeedImageDataLoaderTask?
-        
-        func cancel() {
-            wrapped?.cancel()
-        }
-    }
-    
-    func loadImageData(from url: URL, completion: @escaping (FeedImageDataLoader.Result) -> Void) -> FeedImageDataLoaderTask {
-        let task = TaskWrapper()
-        
-        task.wrapped = primary.loadImageData(from: url) { [weak self] primaryResult in
-            switch primaryResult {
-            case .success:
-                completion(primaryResult)
-                
-            case .failure:
-                task.wrapped = self?.fallback.loadImageData(from: url, completion: completion)
-            }
-        }
-        
-        return task
-    }
-}
+import EssentialApp
 
 final class FeedImageLoaderWithFallbackCompositeTests: XCTestCase {
     func test_loadImageData_doesNotLoadImageDataOnInitialization() {
